+++ Kickstarter-iOS/Views/Controllers/ProjectPamphletViewController.swift
@@ -6,6 +6,9 @@ import UIKit
 public protocol ProjectPamphletViewControllerDelegate: class {
   func projectPamphlet(_ controller: ProjectPamphletViewController,
                        panGestureRecognizerDidChange recognizer: UIPanGestureRecognizer)
+  func projectPamphletViewController(_ projectPamphletViewController: ProjectPamphletViewController,
+                                     shouldTransitionToRewardsWithProject project: Project,
+                                     refTag: RefTag?)
 }
 
 public final class ProjectPamphletViewController: UIViewController {
@@ -202,7 +205,7 @@ public final class ProjectPamphletViewController: UIViewController {
   }
 
   private func goToRewards(project: Project, refTag: RefTag?) {
-
+    self.delegate?.projectPamphletViewController(self, shouldTransitionToRewardsWithProject: project, refTag: refTag)
   }
 
   // MARK: - Private Helpers
