+++ Kickstarter-iOS/Views/Controllers/RewardsCollectionViewControllerTests.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+
+final class RewardsCollectionViewControllerTests: TestCase {
+
+}
