+++ Library/DataSource/UICollectionView-Extensions.swift
@@ -5,6 +5,10 @@ public extension UICollectionView {
     register(cellClass, forCellWithReuseIdentifier: cellClass.description())
   }
 
+  public func register<T: ValueCell>(_ cellClass: T.Type) {
+    self.register(cellClass, forCellWithReuseIdentifier: T.defaultReusableId)
+  }
+
   public func registerCellNibForClass(_ cellClass: AnyClass) {
     let classNameWithoutModule = cellClass
       .description()
